move getEntityFieldID to shared utils.tsx
