+++ incident/src/components/EntityIncidentCard/index.tsx
@@ -13,13 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Entity } from "@backstage/catalog-model";
 import {
   HeaderIconLinkRow,
   IconLinkVerticalProps,
   Progress,
 } from "@backstage/core-components";
-import { ConfigApi, configApiRef, useApi } from "@backstage/core-plugin-api";
+import { configApiRef, useApi } from "@backstage/core-plugin-api";
 import { useEntity } from "@backstage/plugin-catalog-react";
 import {
   Card,
@@ -38,6 +37,7 @@ import { Alert } from "@material-ui/lab";
 import { useState } from "react";
 import { useIncidentList, useIdentity } from "../../hooks/useIncidentRequest";
 import { IncidentListItem } from "../IncidentListItem";
+import { getEntityFieldID } from "../utils";
 
 const IncorrectConfigCard = () => {
   return (
@@ -183,23 +183,4 @@ export const EntityIncidentCard = ({
   );
 };
 
-// Find the ID of the custom field in incident that represents the association
-// to this type of entity.
-//
-// In practice, this will be kind=Component => ID of Affected components field.
-export function getEntityFieldID(config: ConfigApi, entity: Entity) {
-  switch (entity.kind) {
-    case "API":
-      return config.getOptional("incident.fields.api");
-    case "Component":
-      return config.getOptional("incident.fields.component");
-    case "Domain":
-      return config.getOptional("incident.fields.domain");
-    case "System":
-      return config.getOptional("incident.fields.system");
-    case "Group":
-      return config.getOptional("incident.fields.group");
-    default:
-      throw new Error(`unrecognised entity kind: ${entity.kind}`);
-  }
-}
+
